+++ src/components/equipment/Equipment.js
@@ -7,37 +7,48 @@ const Equipment = props => {
     const [equipmentList, setEquipmentList] = useState([])
     const [refresh, setRefresh] = useState(false)
 
-    async function fetchEquipment () {
+    async function fetchEquipment() {
         await ApiManager.getAll("equipments")
-        .then(res => setEquipmentList(res))
+            .then(res => setEquipmentList(res))
     };
 
     const handleDelete = (evt) => {
-        const equipmentToDeleteId = parseInt(evt.target.id.split("--")[1])
-        ApiManager.delete("equipments", equipmentToDeleteId)
-        .then(()=>setRefresh(true))
+        const confirmed = window.confirm("are you sure?")
+        if (confirmed === true) {
+            const equipmentToDeleteId = parseInt(evt.target.id.split("--")[1])
+            ApiManager.delete("equipments", equipmentToDeleteId)
+                .then(() => setRefresh(true))
+        }
     }
 
-    useEffect(()=>{
+    useEffect(() => {
         fetchEquipment();
         setRefresh(false)
-    },[refresh])
+    }, [refresh])
 
     return (
         <>
-        <button
-        id="add-equipment-button"
-        onClick={()=>props.history.push('/equipment/form')}
-        >Add Equipment</button>
-            {equipmentList.map(res => 
-            <>
-            <button
-            id={`delete-equipment-button--${res.id}`}
-            onClick={handleDelete}
-            >delete</button>
-            <EquipmentListItem equipment={res} from={"equipment-page"} {...props} key={res.id} />
-            </>
-            )}
+            <section className="page-container bubble page-margins">
+                <div
+                    id="add-equipment-button"
+                    className="create-button page-margins"
+                    onClick={() => props.history.push('/equipment/form')}
+                >+</div>
+                <div className="bubble psd-heading">Equipment</div>
+                <div className="page-container">
+                    {equipmentList.map(res =>
+                        <>
+
+                            <EquipmentListItem equipment={res} from={"equipment-page"} {...props} key={res.id} handleDelete={handleDelete} />
+                            {/* <div
+                        className="equipment--button"
+                            id={`delete-equipment-button--${res.id}`}
+                            onClick={handleDelete}
+                        >delete</div> */}
+                        </>
+                    )}
+                </div>
+            </section>
         </>
     )
 };
